requirements.txt
fastapi
uvicorn
psycopg2-binary
elasticsearch
python-dotenv
